docs(LICENSE): add a licence
So I'm not expert but I did do some reading around the issue of which
licence to use and decided based on the fact that I don't want FOSS
code I write to be re-packaged and SOLD like it wasn't derived from me or
that the intent wasn't for it to be open. That bothers me.

I don't know if any users will be negatively impacted by this since
unless you intended to sell my code you shouldn't be as far as I know.

Either way if this blocks you somehow open an issue, but as far as I can
see from this: https://www.gnu.org/licenses/gpl-faq.html#InternalDistribution
you can even fork and modify it for internal company usage... you just
can't sell it, which you shouldn't.