feat: restore terminal view on smart toggle (#399)
A continuation of #397 where the idea is taken a bit further, and a complete terminal view is being toggled instead of just the last toggled terminal. When working with the terminals you often want to save your complete view that you've been working on, instead of just closing terminals one by one, and opening them up (again) one by one. There's already a utility mapping that let's you open and close all of your terminals, but this doesn't take into consideration what was previously opened. Therefore to make smart_toggle even smarter we could save the current terminal view, and then open it up the next time it is triggered.

In order to achieve this functionality we could keep a list of terminals that were closed, and use it to go over the list to open them on next smart toggle.