fix: add rate limit for repeated ToggleTerm trigger
Issue: when terminal_mapping is enabled and key-mapping is 'Ctrl + \',
if I hold 'Ctrl + \' for long time, ToggleTerm command will be triggerd
repeatly(show and hide for floating terminal). If the combo-key(Ctrl + \)
is released slightly slower(depend on keyboard), ToggleTerm will be
triggered twice, causing the floating terminal window to briefly appear
and then immediately hide again.

To address the above issue, I have implemented a delay mechanism that
checks the time difference between the current ToggleTerm and the
previous one. If the difference is less than 150ms, it will be ignored.

This modification is quite dirty, and I hope it can serve as a starting
point for further improvements.